terracron/pyproject.toml
[project]
name = "terracron"
version = "0.1.0"
description = "A simple cron job scheduler for Terraform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "schedule>=1.2.2",
]
